集群统计功能Metrics

+++ src/domain/dto/cluter.info.node.dto.go
@@ -4,7 +4,7 @@ type NodeStatus struct {
 	CPU     float64 `json:"cpu"`
 	Memory  float64 `json:"memory"`
 	Storage float64 `json:"storage"`
-	Pods    int     `json:"pods"`
+	Pods    int64   `json:"pods"`
 }
 
 type NodeAddress struct {
